+++ CHANGELOG.md
@@ -63,10 +63,12 @@
 - Add new theme Amber Dusk.
 - Modify list theme.    
 - Add button to open terminal.
-- New version 0.1.9 (2026-02-07)
 
-#### ⟳ Changelog Beta version 0.1.10 (2026-02-08)
+#### ⟳ Changelog Beta version 0.1.9 (2026-02-08)
 - Fixed color of the text in the notes tab.
 - Changed info layout in settings tab.
 - Add copy icon in notes tab.
+- Add support to txt files in notes tab.
+- Add optinon browser in project info tab.
+- Add changes style for buttons in project info tab.